refactor: Judge가 Dealer를 필드로 갖지 않고 매개변수로 받도록 변경
