+++ src/main/java/blackjack/domain/gamer/Dealer.java
@@ -1,4 +1,4 @@
-package blackjack.domain.gamers;
+package blackjack.domain.gamer;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Deck;
