+++ src/main/java/blackjack/domain/gamer/Name.java
@@ -1,4 +1,4 @@
-package blackjack.domain.gamers;
+package blackjack.domain.gamer;
 
 public record Name(String value) {
 }
