+++ src/main/java/blackjack/view/InputView.java
@@ -1,6 +1,6 @@
 package blackjack.view;
 
-import blackjack.domain.gamers.Name;
+import blackjack.domain.gamer.Name;
 import java.util.List;
 import java.util.Scanner;
