+++ src/test/java/blackjack/domain/gamer/PlayerTest.java
@@ -1,11 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.gamer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Rank;
 import blackjack.domain.card.Suit;
-import blackjack.domain.gamers.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
