+++ src/test/java/blackjack/domain/result/JudgeTest.java
@@ -1,11 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.gamers.Dealer;
-import blackjack.domain.gamers.Player;
-import blackjack.domain.result.Judge;
-import blackjack.domain.result.PlayerOutcome;
+import blackjack.domain.TestDeckFactory;
+import blackjack.domain.gamer.Dealer;
+import blackjack.domain.gamer.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -33,9 +32,8 @@ class JudgeTest {
         dealer.draw();
         dealer.draw();
         dealer.draw();
-        final Judge judge = new Judge(dealer);
 
-        final PlayerOutcome playerOutcome = judge.calculatePlayerOutcome(player);
+        final PlayerOutcome playerOutcome = Judge.calculatePlayerOutcome(dealer, player);
 
         assertThat(playerOutcome).isEqualTo(PlayerOutcome.LOSE);
     }
@@ -49,9 +47,8 @@ class JudgeTest {
         player.draw(dealer.drawPlayerCard());
         player.draw(dealer.drawPlayerCard());
         dealer.draw();
-        final Judge judge = new Judge(dealer);
 
-        final PlayerOutcome playerOutcome = judge.calculatePlayerOutcome(player);
+        final PlayerOutcome playerOutcome = Judge.calculatePlayerOutcome(dealer, player);
 
         assertThat(playerOutcome).isEqualTo(PlayerOutcome.LOSE);
     }
@@ -65,9 +62,8 @@ class JudgeTest {
         dealer.draw();
         dealer.draw();
         dealer.draw();
-        final Judge judge = new Judge(dealer);
 
-        final PlayerOutcome playerOutcome = judge.calculatePlayerOutcome(player);
+        final PlayerOutcome playerOutcome = Judge.calculatePlayerOutcome(dealer, player);
 
         assertThat(playerOutcome).isEqualTo(PlayerOutcome.NORMAL_WIN);
     }
@@ -84,9 +80,8 @@ class JudgeTest {
         }
         player.draw(dealer.drawPlayerCard());
         dealer.draw();
-        final Judge judge = new Judge(dealer);
 
-        final PlayerOutcome playerOutcome = judge.calculatePlayerOutcome(player);
+        final PlayerOutcome playerOutcome = Judge.calculatePlayerOutcome(dealer, player);
 
         assertThat(playerOutcome).isEqualTo(PlayerOutcome.PUSH);
     }
@@ -102,9 +97,8 @@ class JudgeTest {
             dealer.drawPlayerCard();
         }
         player.draw(dealer.drawPlayerCard());
-        final Judge judge = new Judge(dealer);
 
-        final PlayerOutcome playerOutcome = judge.calculatePlayerOutcome(player);
+        final PlayerOutcome playerOutcome = Judge.calculatePlayerOutcome(dealer, player);
 
         assertThat(playerOutcome).isEqualTo(PlayerOutcome.BLACKJACK_WIN);
     }
@@ -120,9 +114,8 @@ class JudgeTest {
             dealer.drawPlayerCard();
         }
         dealer.draw();
-        final Judge judge = new Judge(dealer);
 
-        final PlayerOutcome playerOutcome = judge.calculatePlayerOutcome(player);
+        final PlayerOutcome playerOutcome = Judge.calculatePlayerOutcome(dealer, player);
 
         assertThat(playerOutcome).isEqualTo(PlayerOutcome.LOSE);
     }
@@ -135,9 +128,8 @@ class JudgeTest {
         player.draw(dealer.drawPlayerCard());
         player.draw(dealer.drawPlayerCard());
         dealer.draw();
-        final Judge judge = new Judge(dealer);
 
-        final PlayerOutcome playerOutcome = judge.calculatePlayerOutcome(player);
+        final PlayerOutcome playerOutcome = Judge.calculatePlayerOutcome(dealer, player);
 
         assertThat(playerOutcome).isEqualTo(PlayerOutcome.NORMAL_WIN);
     }
@@ -150,9 +142,8 @@ class JudgeTest {
         player.draw(dealer.drawPlayerCard());
         dealer.draw();
         dealer.draw();
-        final Judge judge = new Judge(dealer);
 
-        final PlayerOutcome playerOutcome = judge.calculatePlayerOutcome(player);
+        final PlayerOutcome playerOutcome = Judge.calculatePlayerOutcome(dealer, player);
 
         assertThat(playerOutcome).isEqualTo(PlayerOutcome.LOSE);
     }
@@ -166,9 +157,8 @@ class JudgeTest {
         player.draw(dealer.drawPlayerCard());
         dealer.draw();
         dealer.draw();
-        final Judge judge = new Judge(dealer);
 
-        final PlayerOutcome playerOutcome = judge.calculatePlayerOutcome(player);
+        final PlayerOutcome playerOutcome = Judge.calculatePlayerOutcome(dealer, player);
 
         assertThat(playerOutcome).isEqualTo(PlayerOutcome.PUSH);
     }
